Use a function as the handler instead of a custom async consumer
